requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bcrypt==4.3.0
bidict==0.23.1
certifi==2025.6.15
click==8.2.1
colorama==0.4.6
deprecation==2.1.0
fastapi==0.115.13
frozenlist==1.7.0
gotrue==2.12.0
granian==2.3.4
greenlet==3.2.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.5.0
packaging==25.0
platformdirs==4.3.8
pluggy==1.6.0
postgrest==1.0.2
propcache==0.3.2
psutil==7.0.0
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pytest==8.4.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-engineio==4.12.2
python-multipart==0.0.20
python-socketio==5.13.0
PyYAML==6.0.2
realtime==2.4.3
redis==6.2.0
reflex-hosting-cli==0.1.50
rich==14.0.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlmodel==0.0.24
starlette==0.46.2
storage3==0.11.3
StrEnum==0.4.15
supafunc==0.9.4
tabulate==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
watchfiles==1.1.0
websockets==14.2
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.1
